capsules/web-context/requirements.txt
openai>=1.12.0
pyyaml>=6.0.1
jsonschema>=4.19.0
requests>=2.31.0
ddgs>=1.0.0
beautifulsoup4>=4.12.0
markdownify>=0.11.6
